+++ src-docs/src/views/tables/tables_example.js
@@ -10,6 +10,7 @@ import { section as basicSection } from './basic';
 import { section as paginatedSection } from './paginated';
 import { section as sortingSection } from './sorting';
 import { section as selectionSection } from './selection';
+import { section as footerSection } from './footer';
 import { section as expandingRowsSection } from './expanding_rows';
 import { section as actionsSection } from './actions';
 import {
@@ -48,6 +49,7 @@ export const TableExample = {
     paginatedSection,
     sortingSection,
     selectionSection,
+    footerSection,
     expandingRowsSection,
     actionsSection,
     inMemorySection,